requirements.txt
fastapi~=0.115.11
uvicorn[standard]~=0.34.0
asyncpg~=0.30.0
SQLAlchemy>=2.0
alembic~=1.15.2
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.4.0
python-multipart~=0.0.20
python-dotenv~=1.0.1
bcrypt~=4.3.0
h11~=0.14.0
pip~=25.1
attrs~=25.3.0
wheel~=0.45.1
PySocks~=1.7.1
cryptography~=44.0.2
rsa~=4.9
psycopg2-binary~=2.9.10
annotated-types~=0.7.0
rich~=13.9.4
itsdangerous~=2.2.0
Jinja2~=3.1.6
httpx~=0.28.1
sniffio~=1.3.1
trio~=0.29.0
outcome~=1.3.0.post0
typing_extensions~=4.12.2
six~=1.17.0
click~=8.1.8
uvloop~=0.21.0
starlette~=0.46.1
pydantic~=2.10.6
email_validator~=2.2.0
pydantic_core~=2.27.2
requests~=2.32.3
websockets~=15.0.1
anyio~=4.8.0
httptools~=0.6.4
wsproto~=1.2.0
watchfiles~=1.0.4
gigachat
aiohttp
greenlet
pyjwt
pydantic-settings
pathlib
selenium